EEPROM - do not require an address for commands
Commands like ON and OFF should (usually) have a module address; however, this is not a requirement of X-10
protocol (not specifying an address means previously-addressed modules will react - or none will, if ALL_OFF,
which unaddressed all units was the last command).

This allows to work around an issue when some SMARTHOME modules will not react to commands sent out by CM11a
eeprom - when module won't react to 'ON A1', sending 'ON A1; ON A' will usually make it work.